refact: axios를 fetch로 변경(#8)
