--- src/pages/login/login.module.scss
@@ -1,14 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 100px auto 0;
-  gap: 30px;
-
-  .form {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-  }
-}
